requirements.txt
webdavclient3 
pydantic 
pillow 
pillow-heif 
hachoir

